requirements.txt
selenium 
webdriver-manager
python-dotenv
schedule